feat: extracted logic and clean up
